Built a basic perception of the "home" screen where the grades are displayed: thanks to this guy https://www.youtube.com/watch?v=18VcnYN5_LM. Concept needs some troubleshooting such as:
Description for the description: Believe it or not, numbers alone dont give enough context

Gap between the header and the recycler view: Maybe fill up with adds XDDDD

Not every screen needs a grade list. So another bug to fix

Green checkmarks need to be in the home style of the app.